+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 
 # CHANGELOG
 
+## v1.17.0
+- Add support for Running Train (beta)
+- Add simple map keybinding form
+
 ## v1.16.1
 - Use HID path matching to optionally find serial